PDocker/backend/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
psycopg2-binary
flask-cors
